+++ modules/client/examples/03-encoders/10-remove-address-list-members.ts
@@ -2,7 +2,7 @@
 ### Remove Members (AddressList)
 */
 import {
-  ClientAddressList,
+  AddresslistVotingClient,
   Context,
   ContextPlugin,
 } from "@aragon/sdk-client";
@@ -10,7 +10,7 @@ import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
-const client = new ClientAddressList(contextPlugin);
+const client = new AddresslistVotingClient(contextPlugin);
 
 const members: string[] = [
   "0x1357924680135792468013579246801357924680",