+++ modules/client/examples/04-decoders/07-update-plugin-settings-address-list.ts
@@ -2,7 +2,7 @@
 ### Decode Update Plugin Settings Action (Address List)
 */
 import {
-  ClientAddressList,
+  AddresslistVotingClient,
   Context,
   ContextPlugin,
   VotingSettings,
@@ -11,7 +11,7 @@ import { contextParams } from "../00-client/00-context";
 const context: Context = new Context(contextParams);
 // Create a plugin context from the simple context
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
-const clientAddressList = new ClientAddressList(contextPlugin);
+const clientAddressList = new AddresslistVotingClient(contextPlugin);
 const data: Uint8Array = new Uint8Array([12, 56]);
 
 const params: VotingSettings = clientAddressList.decoding